backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
google-generativeai
pydantic
pandas
